rename files, delete comments
